Camera Improvements, BountyTower file for code seperation, Menu button calls tower build code
