+++ src/Camera.h
@@ -1,0 +1,11 @@
+#ifndef CAMERA_H_INCLUDED
+#define CAMERA_H_INCLUDED
+
+#include "globalvars.h"
+#include "InputState.h"
+#include "util.h"
+#include "Tiles.h"
+
+void cameraControls();
+
+#endif // CAMERA_H_INCLUDED
